+++ docs/CHANGELOG.md
@@ -9,3 +9,4 @@
 - Fix github actions
 - Add TARGET landcover production
 - Force TARGET ROOF fraction to 0.75 when BUILDING fraction is greater than 0.75
+- Force TARGET W indicator to the grid resolution